// projects/ngx-td/src/lib/avatars/models/avatar.interface.ts
/**
 * @description Interface for the `NgxSgAvatarComponent`.
 *
 * @link [Documentation](https://karma-solutions-org.github.io/ngx-sg/docs/avatars#avatarinterface)
 */
export interface AvatarInterface {
  username: string;
  dimension?: number;
  fontSize?: number;
  backgroundColor?: string;
  textColor?: string;
}